chore:add sql

+++ migrations/00005_insert_user_role_value.sql
@@ -4,8 +4,10 @@ INSERT INTO
         `role_name`,
         `role_description`
     )
-VALUES ('admin', '管理员'),
-    ('user', '普通用户');
+VALUES ('admin', 'Administrator'),
+    ('user', 'Regular user')
+ON DUPLICATE KEY UPDATE
+    role_description = VALUES(role_description);
 
 -- +goose Down
-DELETE FROM user_role
+DELETE FROM user_role WHERE role_name IN ('admin', 'user');
